Updates to the chassis
This update to the chassis includes fixations for various parts as well as a part for the electronics to go on, and a servo holder for the front servo.
With this update, the robot can be built and be functional. Although a cover for the electronics is missing, this doesn't affect R.Ian functionality.